refactor(server): extract mesh-request, telemetry, connection, device & status route groups (Refs #3502)
Continues the server.ts route-extraction effort (#3502) by moving 21
inline apiRouter handlers into five cohesive routes/ modules, matching
the established Router-per-module pattern (PRs #3519, #3523).

Extracted (behaviour-preserving, middleware order identical):
- meshRequestRoutes.ts  — /traceroute, /position/request,
  /nodeinfo/request, /neighborinfo/request, /telemetry/request
  (carries the neighbor-info rate-limit map)
- telemetryRoutes.ts    — /direct-neighbors and the /telemetry/* query
  endpoints (:nodeId, /rates, /smarthops, /linkquality,
  DELETE :nodeId/:type, /available/nodes)
- connectionRoutes.ts   — /connection (GET, /disconnect, /reconnect,
  /info, /configure)
- deviceStatusRoutes.ts — /device/tx-status, /device/security-keys
- statusRoutes.ts       — /virtual-node/status, /automation/airtime-status

Mounted connectionRoutes at /connection; the mixed-prefix routers mount
at / to keep full paths verbatim. server.ts inline handler count drops
from 130 to 109. All handler dependencies were already imported from
utils/ modules, so no server.ts-local helpers needed exporting. Added a
*.test.ts per new module (33 new tests).
