+++ src/server/server.ts
@@ -983,6 +983,11 @@ import ignoredNodeRoutes from './routes/ignoredNodeRoutes.js';
 import announceRoutes from './routes/announceRoutes.js';
 import { backupRouter, systemBackupRouter } from './routes/backupRoutes.js';
 import { pushRouter, appriseRouter } from './routes/notificationRoutes.js';
+import meshRequestRoutes from './routes/meshRequestRoutes.js';
+import telemetryRoutes from './routes/telemetryRoutes.js';
+import connectionRoutes from './routes/connectionRoutes.js';
+import deviceStatusRoutes from './routes/deviceStatusRoutes.js';
+import statusRoutes from './routes/statusRoutes.js';
 
 // CSRF token endpoint (must be before CSRF protection middleware)
 apiRouter.get('/csrf-token', csrfTokenEndpoint);
@@ -1111,6 +1116,13 @@ apiRouter.use('/backup', backupRouter);
 apiRouter.use('/system/backup', systemBackupRouter);
 apiRouter.use('/push', pushRouter);
 apiRouter.use('/apprise', appriseRouter);
+apiRouter.use('/connection', connectionRoutes);
+// Mounted at '/' because these routers contain mixed top-level paths
+// (e.g. /traceroute, /telemetry/:nodeId, /device/tx-status, /virtual-node/status).
+apiRouter.use('/', meshRequestRoutes);
+apiRouter.use('/', telemetryRoutes);
+apiRouter.use('/', deviceStatusRoutes);
+apiRouter.use('/', statusRoutes);
 
 // Wire up side-effect callbacks for settingsRoutes
 setSettingsCallbacks({
@@ -2663,44 +2675,6 @@ apiRouter.get('/messages/unread-counts', optionalAuth(), async (req, res) => {
   }
 });
 
-// Get Virtual Node server status per source (requires authentication)
-apiRouter.get('/virtual-node/status', requireAuth(), (_req, res) => {
-  try {
-    const managers = sourceManagerRegistry.getAllManagers() as any[];
-    const sources = managers.map((mgr) => {
-      const vn = mgr.virtualNodeServer;
-      const status = mgr.getStatus?.();
-      const sourceId = status?.sourceId ?? mgr.sourceId;
-      const sourceName = status?.sourceName ?? sourceId;
-      if (!vn) {
-        return {
-          sourceId,
-          sourceName,
-          enabled: false,
-          isRunning: false,
-          allowAdminCommands: false,
-          clientCount: 0,
-          clients: [],
-        };
-      }
-      return {
-        sourceId,
-        sourceName,
-        enabled: true,
-        isRunning: vn.isRunning(),
-        allowAdminCommands: vn.isAdminCommandsAllowed(),
-        clientCount: vn.getClientCount(),
-        clients: vn.getClientDetails(),
-      };
-    });
-
-    res.json({ sources });
-  } catch (error) {
-    logger.error('Error getting virtual node status:', error);
-    res.status(500).json({ error: 'Failed to get virtual node status' });
-  }
-});
-
 // MM-SEC-6: legacy `/api/channels/debug` removed.
 // The route was a `SELECT *` pass-through gated on the unrelated
 // `messages:read` permission, so any user with `messages:read` (granted to
@@ -3882,695 +3856,6 @@ apiRouter.post('/messages/send', optionalAuth(), async (req, res) => {
   }
 });
 
-// Traceroute endpoint
-apiRouter.post('/traceroute', requirePermission('traceroute', 'write'), async (req, res) => {
-  try {
-    const { destination, sourceId: traceSourceId } = req.body;
-    if (!destination) {
-      return res.status(400).json({ error: 'Destination node number is required' });
-    }
-
-    const destinationNum = await parseDestinationNum(destination, traceSourceId, databaseService);
-    if (destinationNum === null) {
-      return res.status(400).json({ error: `Invalid destination: ${destination}` });
-    }
-
-    // Look up the node to get its channel — scope to this source so the channel
-    // reflects the mesh this traceroute will actually traverse.
-    const node = await databaseService.nodes.getNode(destinationNum, traceSourceId);
-    const channel = node?.channel ?? 0; // Default to 0 if node not found or channel not set
-
-    const traceManager = (resolveSourceManager(traceSourceId));
-    await traceManager.sendTraceroute(destinationNum, channel);
-    res.json({
-      success: true,
-      message: `Traceroute request sent to ${destinationNum.toString(16)} on channel ${channel}`,
-    });
-  } catch (error) {
-    logger.error('Error sending traceroute:', error);
-    res.status(500).json({ error: 'Failed to send traceroute' });
-  }
-});
-
-// Position request endpoint
-apiRouter.post('/position/request', requirePermission('messages', 'write'), async (req, res) => {
-  try {
-    const { destination, sourceId: posSourceId } = req.body;
-    if (!destination) {
-      return res.status(400).json({ error: 'Destination node number is required' });
-    }
-
-    const destinationNum = await parseDestinationNum(destination, posSourceId, databaseService);
-    if (destinationNum === null) {
-      return res.status(400).json({ error: `Invalid destination: ${destination}` });
-    }
-
-    // Look up the node to get its channel (scoped to this source)
-    const node = await databaseService.nodes.getNode(destinationNum, posSourceId);
-    // Use explicit channel from request if provided and valid (0-7), otherwise fall back to node's stored channel
-    const channel = (typeof req.body.channel === 'number' && req.body.channel >= 0 && req.body.channel <= 7)
-      ? req.body.channel
-      : (node?.channel ?? 0);
-
-    const posManager = (resolveSourceManager(posSourceId));
-    const { packetId, requestId } = await posManager.sendPositionRequest(destinationNum, channel);
-
-    // Get local node info to create system message
-    const localNodeInfo = posManager.getLocalNodeInfo();
-    logger.info(
-      `📍 localNodeInfo for system message: ${
-        localNodeInfo ? `nodeId=${localNodeInfo.nodeId}, nodeNum=${localNodeInfo.nodeNum}` : 'NULL'
-      }`
-    );
-
-    const isBroadcast = destinationNum === 0xFFFFFFFF;
-
-    if (localNodeInfo) {
-      // Create a system message to record the position request using the actual packet ID and requestId
-      const messageId = `${packetId}`;
-      const timestamp = Date.now();
-
-      // For DMs (channel 0), store as channel -1 to show in DM conversation
-      const messageChannel = channel === 0 ? -1 : channel;
-
-      logger.info(
-        `📍 Inserting position request system message to database: ${messageId} (channel: ${messageChannel}, packetId: ${packetId}, requestId: ${requestId}, broadcast: ${isBroadcast})`
-      );
-      await databaseService.messages.insertMessage({
-        id: messageId,
-        fromNodeNum: localNodeInfo.nodeNum,
-        toNodeNum: destinationNum,
-        fromNodeId: localNodeInfo.nodeId,
-        toNodeId: `!${destinationNum.toString(16).padStart(8, '0')}`,
-        text: isBroadcast ? 'Position broadcast sent' : 'Position exchange requested',
-        channel: messageChannel,
-        portnum: PortNum.TEXT_MESSAGE_APP, // Shows in DM view (DM filter requires TEXT_MESSAGE_APP)
-        // Broadcast packets don't get ACKed, so omit requestId to avoid permanent pending state
-        ...(isBroadcast ? {} : { requestId: requestId }),
-        timestamp: timestamp,
-        rxTime: timestamp,
-        createdAt: timestamp,
-        sourceIp: req.ip ?? null,
-        sourcePath: 'http_api',
-      });
-      logger.info(`📍 Position request system message inserted successfully`);
-    } else {
-      logger.warn(`⚠️ Could not create system message for position request - localNodeInfo is null`);
-    }
-
-    res.json({
-      success: true,
-      message: `Position request sent to ${destinationNum.toString(16)} on channel ${channel}`,
-    });
-  } catch (error) {
-    logger.error('Error sending position request:', error);
-    res.status(500).json({ error: 'Failed to send position request' });
-  }
-});
-
-// NodeInfo request endpoint (Exchange Node Info - triggers key exchange)
-apiRouter.post('/nodeinfo/request', requirePermission('messages', 'write'), async (req, res) => {
-  try {
-    const { destination, sourceId: niSourceId } = req.body;
-    if (!destination) {
-      return res.status(400).json({ error: 'Destination node number is required' });
-    }
-
-    const destinationNum = await parseDestinationNum(destination, niSourceId, databaseService);
-    if (destinationNum === null) {
-      return res.status(400).json({ error: `Invalid destination: ${destination}` });
-    }
-
-    // Look up the node to get its channel (scoped to this source)
-    const node = await databaseService.nodes.getNode(destinationNum, niSourceId);
-    const channel = node?.channel ?? 0; // Default to 0 if node not found or channel not set
-
-    const niManager = (resolveSourceManager(niSourceId));
-    const { packetId, requestId } = await niManager.sendNodeInfoRequest(destinationNum, channel);
-
-    // Get local node info to create system message
-    const localNodeInfo = niManager.getLocalNodeInfo();
-    logger.info(
-      `📇 localNodeInfo for system message: ${
-        localNodeInfo ? `nodeId=${localNodeInfo.nodeId}, nodeNum=${localNodeInfo.nodeNum}` : 'NULL'
-      }`
-    );
-
-    if (localNodeInfo) {
-      // Create a system message to record the nodeinfo request using the actual packet ID and requestId
-      const messageId = `${packetId}`;
-      const timestamp = Date.now();
-
-      // For DMs (channel 0), store as channel -1 to show in DM conversation
-      const messageChannel = channel === 0 ? -1 : channel;
-
-      logger.info(
-        `📇 Inserting nodeinfo request system message to database: ${messageId} (channel: ${messageChannel}, packetId: ${packetId}, requestId: ${requestId})`
-      );
-      await databaseService.messages.insertMessage({
-        id: messageId,
-        fromNodeNum: localNodeInfo.nodeNum,
-        toNodeNum: destinationNum,
-        fromNodeId: localNodeInfo.nodeId,
-        toNodeId: `!${destinationNum.toString(16).padStart(8, '0')}`,
-        text: 'User info exchange requested',
-        channel: messageChannel,
-        portnum: PortNum.TEXT_MESSAGE_APP, // Shows in DM view (DM filter requires TEXT_MESSAGE_APP)
-        requestId: requestId, // Store requestId for ACK matching
-        timestamp: timestamp,
-        rxTime: timestamp,
-        createdAt: timestamp,
-        sourceIp: req.ip ?? null,
-        sourcePath: 'http_api',
-      });
-      logger.info(`📇 NodeInfo request system message inserted successfully`);
-    } else {
-      logger.warn(`⚠️ Could not create system message for nodeinfo request - localNodeInfo is null`);
-    }
-
-    res.json({
-      success: true,
-      message: `NodeInfo request sent to ${destinationNum.toString(16)} on channel ${channel}`,
-    });
-  } catch (error) {
-    logger.error('Error sending nodeinfo request:', error);
-    res.status(500).json({ error: 'Failed to send nodeinfo request' });
-  }
-});
-
-// NeighborInfo request endpoint (request neighbor info from remote node)
-// Rate limit: one request per destination every 180 seconds (firmware limit is ~3 minutes)
-const neighborInfoRequestTimestamps = new Map<number, number>();
-const NEIGHBOR_INFO_RATE_LIMIT_MS = 180_000;
-
-apiRouter.post('/neighborinfo/request', requirePermission('traceroute', 'write'), async (req, res) => {
-  try {
-    const { destination } = req.body;
-    if (!destination) {
-      return res.status(400).json({ error: 'Destination node number is required' });
-    }
-
-    const { sourceId: neighborSourceId } = req.body;
-    const destinationNum = await parseDestinationNum(destination, neighborSourceId, databaseService);
-    if (destinationNum === null) {
-      return res.status(400).json({ error: `Invalid destination: ${destination}` });
-    }
-
-    // Eligibility check: only allow requests to local node or 0-hop nodes
-    const neighborManager = (resolveSourceManager(neighborSourceId));
-    const localNodeNum = neighborManager.getLocalNodeInfo()?.nodeNum;
-    // Scope to the target source so hopsAway/channel reflect this mesh
-    const node = await databaseService.nodes.getNode(destinationNum, neighborSourceId);
-    const isLocalNode = localNodeNum != null && Number(destinationNum) === Number(localNodeNum);
-    const isDirectNode = node != null && node.hopsAway != null && Number(node.hopsAway) === 0;
-
-    if (!isLocalNode && !isDirectNode) {
-      return res.status(403).json({
-        error: 'Neighbor info requests are only allowed for the local node or directly-heard (0-hop) nodes',
-        eligible: false,
-      });
-    }
-
-    // Rate limiting per destination
-    const lastRequest = neighborInfoRequestTimestamps.get(Number(destinationNum));
-    const now = Date.now();
-    if (lastRequest) {
-      if ((now - lastRequest) < NEIGHBOR_INFO_RATE_LIMIT_MS) {
-        const retryAfter = Math.ceil((NEIGHBOR_INFO_RATE_LIMIT_MS - (now - lastRequest)) / 1000);
-        return res.status(429).json({
-          error: 'Rate limited: firmware limits neighbor info responses to once per 3 minutes',
-          retryAfter,
-        });
-      }
-      // Expired entry — clean up
-      neighborInfoRequestTimestamps.delete(Number(destinationNum));
-    }
-
-    const channel = node?.channel ?? 0; // Default to 0 if node not found or channel not set
-
-    const { packetId, requestId } = await neighborManager.sendNeighborInfoRequest(destinationNum, channel);
-    neighborInfoRequestTimestamps.set(Number(destinationNum), now);
-
-    logger.info(`🏠 NeighborInfo request sent to ${destinationNum.toString(16)} on channel ${channel}, packetId=${packetId}, requestId=${requestId}`);
-
-    res.json({
-      success: true,
-      message: `NeighborInfo request sent to ${destinationNum.toString(16)} on channel ${channel}`,
-      packetId,
-      requestId
-    });
-  } catch (error) {
-    logger.error('Error sending neighborinfo request:', error);
-    res.status(500).json({ error: 'Failed to send neighborinfo request' });
-  }
-});
-
-// Telemetry request endpoint (request telemetry from remote node)
-apiRouter.post('/telemetry/request', requirePermission('messages', 'write'), async (req, res) => {
-  try {
-    const { destination, telemetryType, sourceId: telSourceId } = req.body;
-    if (!destination) {
-      return res.status(400).json({ error: 'Destination node number is required' });
-    }
-
-    // Validate telemetry type if provided
-    const validTypes = ['device', 'environment', 'airQuality', 'power'];
-    if (telemetryType && !validTypes.includes(telemetryType)) {
-      return res.status(400).json({ error: `Invalid telemetry type. Must be one of: ${validTypes.join(', ')}` });
-    }
-
-    const destinationNum = await parseDestinationNum(destination, telSourceId, databaseService);
-    if (destinationNum === null) {
-      return res.status(400).json({ error: `Invalid destination: ${destination}` });
-    }
-
-    // Look up the node to get its channel (scoped to this source)
-    const node = await databaseService.nodes.getNode(destinationNum, telSourceId);
-    const channel = node?.channel ?? 0; // Default to 0 if node not found or channel not set
-
-    const telManager = (resolveSourceManager(telSourceId));
-    const { packetId, requestId } = await telManager.sendTelemetryRequest(
-      destinationNum,
-      channel,
-      telemetryType as 'device' | 'environment' | 'airQuality' | 'power' | undefined
-    );
-
-    const typeLabel = telemetryType || 'device';
-    logger.info(`📊 Telemetry request (${typeLabel}) sent to ${destinationNum.toString(16)} on channel ${channel}, packetId=${packetId}, requestId=${requestId}`);
-
-    res.json({
-      success: true,
-      message: `Telemetry request (${typeLabel}) sent to ${destinationNum.toString(16)} on channel ${channel}`,
-      packetId,
-      requestId
-    });
-  } catch (error) {
-    logger.error('Error sending telemetry request:', error);
-    res.status(500).json({ error: 'Failed to send telemetry request' });
-  }
-});
-
-// Get direct neighbor RSSI statistics from zero-hop packets
-// This helps identify which nodes we've heard directly (no relays)
-apiRouter.get('/direct-neighbors', requirePermission('info', 'read'), async (req, res) => {
-  try {
-    const hoursBack = parseInt(req.query.hours as string) || 24;
-    const stats = await databaseService.getDirectNeighborStatsAsync(hoursBack);
-
-    res.json({
-      success: true,
-      data: stats,
-      count: Object.keys(stats).length
-    });
-  } catch (error) {
-    logger.error('Error getting direct neighbor stats:', error);
-    res.status(500).json({ error: 'Failed to fetch direct neighbor statistics' });
-  }
-});
-
-// Get telemetry data for a node
-apiRouter.get('/telemetry/:nodeId', optionalAuth(), async (req, res) => {
-  try {
-    // Allow users with info read OR dashboard read (dashboard needs telemetry data)
-    if (
-      !req.user?.isAdmin &&
-      !(req.user ? await hasPermission(req.user, 'info', 'read') : false) &&
-      !(req.user ? await hasPermission(req.user, 'dashboard', 'read') : false)
-    ) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-
-    const { nodeId } = req.params;
-
-    // Check channel-based access for this node
-    if (!await checkNodeChannelAccess(nodeId, req.user)) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-    // parseFloat (not parseInt) so sub-hour windows like 0.25h (15 minutes)
-    // from the Device Info time-range selector survive the round-trip.
-    const rawHours = req.query.hours ? parseFloat(req.query.hours as string) : 24;
-    const hoursParam = Number.isFinite(rawHours) && rawHours > 0 ? rawHours : 24;
-    const telSourceId = req.query.sourceId as string | undefined;
-
-    // Calculate cutoff timestamp for filtering
-    const cutoffTime = Date.now() - hoursParam * 60 * 60 * 1000;
-
-    // Check if node has private position override
-    const nodeNum = parseInt(nodeId.replace('!', ''), 16);
-    const node = await databaseService.nodes.getNode(nodeNum, telSourceId);
-    const isPrivate = node?.positionOverrideIsPrivate === true;
-    const canViewPrivate = !!req.user && await hasPermission(req.user, 'nodes_private', 'read');
-
-    // Use the averaged query for graph data on every backend (SQLite,
-    // PostgreSQL, MySQL). The interval is chosen dynamically to target a
-    // manageable, roughly fixed number of points per telemetry type, so short
-    // windows keep near-full resolution while long windows return the full
-    // history downsampled instead of being truncated to the newest N rows.
-    const telemetry = await databaseService.getTelemetryByNodeAveragedAsync(nodeId, cutoffTime, undefined, hoursParam, telSourceId);
-
-    // Filter out location telemetry if private and unauthorized
-    let processedTelemetry = telemetry;
-    if (isPrivate && !canViewPrivate) {
-      processedTelemetry = telemetry.filter(t =>
-        !['latitude', 'longitude', 'altitude'].includes(t.telemetryType)
-      );
-    }
-
-    res.json(processedTelemetry);
-  } catch (error) {
-    logger.error('Error fetching telemetry:', error);
-    res.status(500).json({ error: 'Failed to fetch telemetry' });
-  }
-});
-
-// Get packet rate statistics (packets per minute) for a node
-apiRouter.get('/telemetry/:nodeId/rates', optionalAuth(), async (req, res) => {
-  try {
-    // Allow users with info read OR dashboard read
-    if (
-      !req.user?.isAdmin &&
-      !(req.user ? await hasPermission(req.user, 'info', 'read') : false) &&
-      !(req.user ? await hasPermission(req.user, 'dashboard', 'read') : false)
-    ) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-
-    const { nodeId } = req.params;
-
-    // Check channel-based access for this node
-    if (!await checkNodeChannelAccess(nodeId, req.user)) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-    const hoursParam = req.query.hours ? parseInt(req.query.hours as string) : 24;
-    const ratesSourceId = req.query.sourceId as string | undefined;
-
-    // Calculate cutoff timestamp for filtering
-    const cutoffTime = Date.now() - hoursParam * 60 * 60 * 1000;
-
-    // The 7 packet statistics types we want rates for
-    const packetTypes = [
-      'numPacketsRx',
-      'numPacketsRxBad',
-      'numRxDupe',
-      'numPacketsTx',
-      'numTxDropped',
-      'numTxRelay',
-      'numTxRelayCanceled',
-    ];
-
-    let rates: Record<string, Array<{ timestamp: number; ratePerMinute: number }>>;
-
-    // For PostgreSQL/MySQL, calculate rates from raw telemetry
-    if (databaseService.drizzleDbType === 'postgres' || databaseService.drizzleDbType === 'mysql') {
-      rates = {};
-      for (const type of packetTypes) {
-        rates[type] = [];
-      }
-
-      // Fetch telemetry for each packet type and calculate rates
-      for (const type of packetTypes) {
-        const telemetry = await databaseService.telemetry.getTelemetryByNode(
-          nodeId, 5000, cutoffTime, undefined, 0, type, ratesSourceId
-        );
-
-        // Sort by timestamp ascending for rate calculation
-        telemetry.sort((a, b) => a.timestamp - b.timestamp);
-
-        // Calculate rates from consecutive samples
-        for (let i = 1; i < telemetry.length; i++) {
-          const prev = telemetry[i - 1];
-          const curr = telemetry[i];
-          const timeDiffMs = curr.timestamp - prev.timestamp;
-          const valueDiff = curr.value - prev.value;
-
-          if (timeDiffMs > 0 && valueDiff >= 0) {
-            const timeDiffMinutes = timeDiffMs / 60000;
-            const ratePerMinute = valueDiff / timeDiffMinutes;
-            rates[type].push({
-              timestamp: curr.timestamp,
-              ratePerMinute: Math.round(ratePerMinute * 100) / 100,
-            });
-          }
-        }
-      }
-    } else {
-      rates = await databaseService.getPacketRatesAsync(nodeId, packetTypes, cutoffTime, ratesSourceId);
-    }
-
-    res.json(rates);
-  } catch (error) {
-    logger.error('Error fetching packet rates:', error);
-    res.status(500).json({ error: 'Failed to fetch packet rates' });
-  }
-});
-
-// Get smart hops statistics (min/max/avg hop counts over time) for a node
-apiRouter.get('/telemetry/:nodeId/smarthops', optionalAuth(), async (req, res) => {
-  try {
-    // Allow users with info read OR dashboard read
-    if (
-      !req.user?.isAdmin &&
-      !(req.user ? await hasPermission(req.user, 'info', 'read') : false) &&
-      !(req.user ? await hasPermission(req.user, 'dashboard', 'read') : false)
-    ) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-
-    const { nodeId } = req.params;
-
-    // Check channel-based access for this node
-    if (!await checkNodeChannelAccess(nodeId, req.user)) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-    // Validate and clamp hours (1-168, default 24)
-    const hoursParam = Math.max(1, Math.min(168, parseInt(req.query.hours as string) || 24));
-    // Validate and clamp interval (5-60 minutes, default 15)
-    const intervalParam = Math.max(5, Math.min(60, parseInt(req.query.interval as string) || 15));
-
-    // Calculate cutoff timestamp for filtering
-    const cutoffTime = Date.now() - hoursParam * 60 * 60 * 1000;
-
-    // Get smart hops statistics
-    const stats = await databaseService.getSmartHopsStatsAsync(nodeId, cutoffTime, intervalParam);
-
-    res.json({ success: true, data: stats });
-  } catch (error) {
-    logger.error('Error fetching smart hops stats:', error);
-    res.status(500).json({ error: 'Failed to fetch smart hops statistics' });
-  }
-});
-
-// Get link quality history for a node
-apiRouter.get('/telemetry/:nodeId/linkquality', optionalAuth(), async (req, res) => {
-  try {
-    // Allow users with info read OR dashboard read
-    if (
-      !req.user?.isAdmin &&
-      !(req.user ? await hasPermission(req.user, 'info', 'read') : false) &&
-      !(req.user ? await hasPermission(req.user, 'dashboard', 'read') : false)
-    ) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-
-    const { nodeId } = req.params;
-
-    // Check channel-based access for this node
-    if (!await checkNodeChannelAccess(nodeId, req.user)) {
-      return res.status(403).json({ error: 'Insufficient permissions' });
-    }
-    // Validate and clamp hours (1-168, default 24)
-    const hoursParam = Math.max(1, Math.min(168, parseInt(req.query.hours as string) || 24));
-
-    // Calculate cutoff timestamp for filtering
-    const cutoffTime = Date.now() - hoursParam * 60 * 60 * 1000;
-
-    // Get link quality history
-    const history = await databaseService.getLinkQualityHistoryAsync(nodeId, cutoffTime);
-
-    res.json({ success: true, data: history });
-  } catch (error) {
-    logger.error('Error fetching link quality history:', error);
-    res.status(500).json({ error: 'Failed to fetch link quality history' });
-  }
-});
-
-// Delete telemetry data for a specific node and type
-apiRouter.delete('/telemetry/:nodeId/:telemetryType', requireAuth(), requirePermission('info', 'write'), async (req, res) => {
-  try {
-    const { nodeId, telemetryType } = req.params;
-
-    logger.info(`Purging telemetry data for node ${nodeId}, type ${telemetryType}`);
-
-    const deleted = await databaseService.telemetry.deleteTelemetryByNodeAndType(nodeId, telemetryType);
-
-    if (deleted) {
-      logger.info(`Successfully purged ${telemetryType} telemetry for node ${nodeId}`);
-      res.json({ success: true, message: `Telemetry data purged successfully` });
-    } else {
-      res.status(404).json({ error: 'No telemetry data found to delete' });
-    }
-  } catch (error) {
-    logger.error('Error purging telemetry data:', error);
-    res.status(500).json({ error: 'Failed to purge telemetry data' });
-  }
-});
-
-// Check which nodes have telemetry data
-apiRouter.get('/telemetry/available/nodes', requirePermission('info', 'read'), async (req, res) => {
-  try {
-    const telAvailSourceId = req.query.sourceId as string | undefined;
-    const allNodes = await databaseService.nodes.getAllNodes(telAvailSourceId);
-    // Filter nodes based on channel read permissions
-    const nodes = await filterNodesByChannelPermission(allNodes, (req as any).user);
-
-    const nodesWithTelemetry: string[] = [];
-    const nodesWithWeather: string[] = [];
-    const nodesWithEstimatedPosition: string[] = [];
-    // Known nodes with neither a real nor an estimated position (issue #3271 map counter).
-    const nodesUnmapped: string[] = [];
-
-    const weatherTypes = new Set(['temperature', 'humidity', 'pressure']);
-
-    // Efficient bulk query: get all telemetry types for all nodes at once
-    const nodeTelemetryTypes = await databaseService.getAllNodesTelemetryTypesAsync(telAvailSourceId);
-    // Global estimated positions (one per physical node, pooled across sources).
-    const estimatedRows = await databaseService.getAllEstimatedPositionsAsync();
-    const estimatedPositionMap = new Map(estimatedRows.map(r => [r.nodeId, r]));
-    const estimatedUncertainty: Record<string, number> = {};
-
-    nodes.forEach(node => {
-      const telemetryTypes = nodeTelemetryTypes.get(node.nodeId);
-      if (telemetryTypes && telemetryTypes.length > 0) {
-        nodesWithTelemetry.push(node.nodeId);
-
-        // Check if any telemetry type is weather-related
-        const hasWeather = telemetryTypes.some(t => weatherTypes.has(t));
-        if (hasWeather) {
-          nodesWithWeather.push(node.nodeId);
-        }
-      }
-
-      // Estimated-position / unmapped status is independent of telemetry presence.
-      // A user-set override counts as a known position — we don't want to draw an
-      // uncertainty circle on a node the user has explicitly placed (issue #2847).
-      const eff = getEffectiveDbNodePosition(node);
-      const hasRealPosition = eff.latitude != null && eff.longitude != null;
-      const estimate = estimatedPositionMap.get(node.nodeId);
-      const hasEstimatedPosition = estimate !== undefined;
-      if (hasEstimatedPosition && !hasRealPosition) {
-        nodesWithEstimatedPosition.push(node.nodeId);
-        if (estimate.uncertaintyKm != null) {
-          estimatedUncertainty[node.nodeId] = estimate.uncertaintyKm;
-        }
-      }
-      if (!hasRealPosition && !hasEstimatedPosition) {
-        nodesUnmapped.push(node.nodeId);
-      }
-    });
-
-    // Check for PKC-enabled nodes
-    const nodesWithPKC: string[] = [];
-
-    // Get the local node ID to ensure it's always marked as secure.
-    // Read per-source so multi-source deployments don't always show the first
-    // source's local node as "secure local node" for every source view.
-    const localNodeNumStr = await databaseService.settings.getSettingForSource(
-      telAvailSourceId ?? null,
-      'localNodeNum'
-    );
-    let localNodeId: string | null = null;
-    if (localNodeNumStr) {
-      const localNodeNum = parseInt(localNodeNumStr, 10);
-      localNodeId = `!${localNodeNum.toString(16).padStart(8, '0')}`;
-    }
-
-    nodes.forEach(node => {
-      // Local node is always secure (direct TCP/serial connection, no mesh encryption needed)
-      // OR node has PKC enabled
-      if (node.nodeId === localNodeId || node.hasPKC || node.publicKey) {
-        nodesWithPKC.push(node.nodeId);
-      }
-    });
-
-    res.json({
-      nodes: nodesWithTelemetry,
-      weather: nodesWithWeather,
-      estimatedPosition: nodesWithEstimatedPosition,
-      estimatedUncertainty,
-      unmapped: nodesUnmapped,
-      unmappedCount: nodesUnmapped.length,
-      pkc: nodesWithPKC,
-    });
-  } catch (error) {
-    logger.error('Error checking telemetry availability:', error);
-    res.status(500).json({ error: 'Failed to check telemetry availability' });
-  }
-});
-
-// Connection status endpoint
-apiRouter.get('/connection', optionalAuth(), async (req, res) => {
-  try {
-    const connSourceId = req.query.sourceId as string | undefined;
-    // When the caller explicitly names a sourceId but no manager is registered
-    // for it (e.g. autoConnect=false, or user manually disconnected via
-    // /api/sources/:id/disconnect — issue #2773), return a stable
-    // "not connected" response instead of silently falling back to the legacy
-    // singleton. The singleton is the primary source's manager and would
-    // otherwise leak its state across sources.
-    if (connSourceId && !sourceManagerRegistry.getManager(connSourceId)) {
-      res.json({
-        connected: false,
-        nodeResponsive: false,
-        configuring: false,
-        userDisconnected: false,
-      });
-      return;
-    }
-    const connManager = resolveSourceManager(connSourceId);
-    const status = await connManager.getConnectionStatus();
-    // Hide nodeIp from anonymous users
-    if (!req.session.userId) {
-      const { nodeIp, ...statusWithoutNodeIp } = status;
-      res.json(statusWithoutNodeIp);
-    } else {
-      res.json(status);
-    }
-  } catch (error) {
-    logger.error('Error getting connection status:', error);
-    res.status(500).json({ error: 'Failed to get connection status' });
-  }
-});
-
-// Check if TX is disabled
-apiRouter.get('/device/tx-status', optionalAuth(), async (req, res) => {
-  try {
-    const txSourceId = req.query.sourceId as string | undefined;
-    const txManager = resolveSourceManager(txSourceId);
-    const deviceConfig = await txManager.getDeviceConfig();
-    const txEnabled = deviceConfig?.lora?.txEnabled !== false; // Default to true if undefined
-    res.json({ txEnabled });
-  } catch (error) {
-    logger.error('Error getting TX status:', error);
-    res.status(500).json({ error: 'Failed to get TX status' });
-  }
-});
-
-// Get security keys (public and private) for the local node.
-// MM-SEC-5: gated on `requireAdmin()` because the response includes the
-// device's PKI private key. Any holder of that key can decrypt PKI DMs the
-// local node receives and forge signed packets from it.
-apiRouter.get('/device/security-keys', requireAdmin(), async (req, res) => {
-  try {
-    const skSourceId = req.query.sourceId as string | undefined;
-    const skManager = resolveSourceManager(skSourceId);
-    const keys = skManager.getSecurityKeys();
-    res.json(keys);
-  } catch (error) {
-    logger.error('Error getting security keys:', error);
-    res.status(500).json({ error: 'Failed to get security keys' });
-  }
-});
-
 /**
  * Resolve the connection host/port for an `/api/poll` or `/api/config` caller
  * scoped to a particular sourceId. Returns the active source's `config.host`
@@ -5055,124 +4340,6 @@ apiRouter.get('/poll', optionalAuth(), async (req, res) => {
   } catch (error) {
     logger.error('Error in consolidated poll endpoint:', error);
     res.status(500).json({ error: 'Failed to fetch polling data' });
-  }
-});
-
-// User-initiated disconnect endpoint
-apiRouter.post('/connection/disconnect', requirePermission('connection', 'write'), async (req, res) => {
-  try {
-    const { sourceId: disconnectSourceId } = req.body;
-    const disconnectManager = (resolveSourceManager(disconnectSourceId));
-    await disconnectManager.userDisconnect();
-
-    // Audit log
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'connection_disconnected',
-      'connection',
-      'User initiated disconnect',
-      req.ip || null
-    );
-
-    res.json({ success: true, status: 'user-disconnected' });
-  } catch (error) {
-    logger.error('Error disconnecting:', error);
-    res.status(500).json({ error: 'Failed to disconnect' });
-  }
-});
-
-// User-initiated reconnect endpoint
-apiRouter.post('/connection/reconnect', requirePermission('connection', 'write'), async (req, res) => {
-  try {
-    const { sourceId: reconnectSourceId } = req.body;
-    const reconnectManager = (resolveSourceManager(reconnectSourceId));
-    const success = await reconnectManager.userReconnect();
-
-    // Audit log
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'connection_reconnected',
-      'connection',
-      JSON.stringify({ success }),
-      req.ip || null
-    );
-
-    res.json({
-      success,
-      status: success ? 'connecting' : 'disconnected',
-    });
-  } catch (error) {
-    logger.error('Error reconnecting:', error);
-    res.status(500).json({ error: 'Failed to reconnect' });
-  }
-});
-
-// Get detailed connection info (authenticated users only)
-apiRouter.get('/connection/info', requireAuth(), async (req, res) => {
-  try {
-    const ciSourceId = req.query.sourceId as string | undefined;
-    const ciManager = resolveSourceManager(ciSourceId);
-    const status = await ciManager.getConnectionStatus();
-    const env = getEnvironmentConfig();
-    const ipOverride = await databaseService.settings.getSetting('meshtasticNodeIpOverride');
-    const portOverride = await databaseService.settings.getSetting('meshtasticTcpPortOverride');
-
-    res.json({
-      ...status,
-      defaultIp: env.meshtasticNodeIp,
-      defaultPort: env.meshtasticTcpPort,
-      isOverridden: !!(ipOverride || portOverride),
-      tcpPort: portOverride ? parseInt(portOverride, 10) : env.meshtasticTcpPort
-    });
-  } catch (error) {
-    logger.error('Error getting connection info:', error);
-    res.status(500).json({ error: 'Failed to get connection info' });
-  }
-});
-
-// Configure connection IP address (admin only)
-apiRouter.post('/connection/configure', requireAdmin(), async (req, res) => {
-  try {
-    const { nodeIp } = req.body;
-
-    // Validate IP format (IPv4 address or hostname, with optional port)
-    // Accepts: 192.168.1.100, 192.168.1.100:4403, hostname, hostname:4403
-    const ipRegex = /^(?:(?:(?:25[0-5]|2[0-4]\d|1\d{2}|[1-9]?\d)\.){3}(?:25[0-5]|2[0-4]\d|1\d{2}|[1-9]?\d)|[\w.-]+)(?::\d{1,5})?$/;
-    if (!nodeIp || !ipRegex.test(nodeIp)) {
-      return res.status(400).json({ error: 'Invalid IP address or hostname' });
-    }
-
-    // Validate port range if specified
-    const portMatch = nodeIp.match(/:(\d+)$/);
-    if (portMatch) {
-      const port = parseInt(portMatch[1], 10);
-      if (port < 1 || port > 65535) {
-        return res.status(400).json({ error: 'Port must be between 1 and 65535' });
-      }
-    }
-
-    // Set the override
-    const { sourceId: connConfigSourceId } = req.body;
-    const connConfigManager = (resolveSourceManager(connConfigSourceId));
-    await connConfigManager.setNodeIpOverride(nodeIp);
-
-    // Audit log
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'connection_address_changed',
-      'connection',
-      JSON.stringify({ address: nodeIp }),
-      req.ip || null
-    );
-
-    res.json({
-      success: true,
-      message: 'Node address updated. Reconnecting...',
-      nodeIp
-    });
-  } catch (error) {
-    logger.error('Error configuring connection:', error);
-    res.status(500).json({ error: 'Failed to configure connection' });
   }
 });
 
@@ -6274,21 +5441,6 @@ apiRouter.post('/user/map-preferences', requireAuth(), async (req, res) => {
   } catch (error) {
     logger.error('Error saving user map preferences:', error);
     res.status(500).json({ error: 'Failed to save map preferences' });
-  }
-});
-
-// Airtime cutoff status — drives the live banner on the Automation page.
-// Returns the configured threshold + measurement source, the effective Channel
-// Utilization (null if unknown), how many neighbours were sampled (neighbours
-// mode), and whether automations are paused.
-apiRouter.get('/automation/airtime-status', requirePermission('automation', 'read'), async (req, res) => {
-  try {
-    const airtimeSourceId = (req.query.sourceId as string) || null;
-    const mgr = resolveSourceManager(airtimeSourceId);
-    res.json(await mgr.getAirtimeCutoffStatus());
-  } catch (error) {
-    logger.error('Error fetching airtime cutoff status:', error);
-    res.status(500).json({ error: 'Failed to fetch airtime cutoff status' });
   }
 });
